share row packing and ts logic between sources and sinks
